+++ piqture/embeddings/amplitude_encoder.py
