+++ piqture/embeddings/image_embedding.py
