+++ piqture/embeddings/image_embeddings/mcrqi.py
@@ -5,7 +5,7 @@ import math
 import numpy as np
 from qiskit.circuit import QuantumCircuit
 from qiskit.circuit.library import MCMT, RYGate
-from piqture.data_encoder.image_embedding import ImageEmbedding
+from piqture.embeddings.image_embedding import ImageEmbedding
 from piqture.mixin.image_embedding_mixin import ImageMixin
