+++ piqture/neural_networks/layers/convolutional_layer.py
@@ -14,7 +14,7 @@ from __future__ import annotations
 from typing import Optional
 from qiskit.circuit import QuantumCircuit
 from piqture.neural_networks.layers.base_layer import BaseLayer
-from piqture.tensor_network_circuits.mera import MERA
+from piqture.tensor_networks.mera import MERA
 
 
 class QuantumConvolutionalLayer(BaseLayer):
