+++ piqture/tensor_networks/ttn.py
@@ -15,7 +15,7 @@ from typing import Callable
 import numpy as np
 from qiskit.circuit import QuantumCircuit, ParameterVector
 from piqture.gates.two_qubit_unitary import TwoQubitUnitary
-from piqture.tensor_network_circuits.base_tensor_network import (
+from piqture.tensor_networks.base_tensor_network import (
     BaseTensorNetwork,
 )
