+++ tests/conftest.py
@@ -11,7 +11,7 @@
 """conftest.py"""
 
 # pylint: disable=unused-import
-from tests.data_encoder.image_representations.test_frqi import (
+from tests.embeddings.image_embeddings.test_frqi import (
     circuit_pixel_position_fixture,
 )
 from tests.tensor_networks.test_mps import parameterization_mapper_fixture
