+++ tests/embeddings/image_embeddings/__init__.py
