+++ tests/embeddings/image_embeddings/test_frqi.py
@@ -18,7 +18,7 @@ import numpy as np
 import pytest
 from pytest import raises
 from qiskit.circuit import QuantumCircuit, ParameterVector
-from piqture.data_encoder.image_representations.frqi import FRQI
+from piqture.embeddings.image_embeddings.frqi import FRQI
 
 PIXEL_POS_BINARY2 = ["00", "01", "10", "11"]
 
@@ -146,7 +146,7 @@ class TestFRQI:
             test_circuit = circuit_pixel_position(img_dims, pixel_pos_binary)
 
             with mock.patch(
-                "piqture.data_encoder.image_representations.frqi.FRQI.circuit",
+                "piqture.embeddings.image_embeddings.frqi.FRQI.circuit",
                 new_callable=lambda: mock_circuit,
             ):
                 frqi_object.pixel_position(pixel_pos_binary)
@@ -166,7 +166,7 @@ class TestFRQI:
             test_circuit = circuit_pixel_value(img_dims, pixel_vals, pixel)
 
             with mock.patch(
-                "piqture.data_encoder.image_representations.frqi.FRQI.circuit",
+                "piqture.embeddings.image_embeddings.frqi.FRQI.circuit",
                 new_callable=lambda: mock_circuit,
             ):
                 frqi_object.pixel_value(pixel_pos=pixel)
@@ -206,7 +206,7 @@ class TestFRQI:
             )
 
         with mock.patch(
-            "piqture.data_encoder.image_representations.frqi.FRQI.circuit",
+            "piqture.embeddings.image_embeddings.frqi.FRQI.circuit",
             new_callable=lambda: mock_circuit,
         ):
             frqi_object.frqi()
