+++ tests/tensor_networks/test_mera.py
@@ -16,7 +16,7 @@ import numpy as np
 import pytest
 from pytest import raises
 from qiskit.circuit import QuantumCircuit, ParameterVector
-from piqture.tensor_network_circuits import MERA
+from piqture.tensor_networks import MERA
 from piqture.gates.two_qubit_unitary import TwoQubitUnitary
 
 
@@ -186,7 +186,7 @@ class TestMERA:
         # pylint: disable=line-too-long
         """Tests the mera_backbone method call via the mera_simple function."""
         with mock.patch(
-            "piqture.tensor_network_circuits.mera.MERA.mera_backbone"
+            "piqture.tensor_networks.mera.MERA.mera_backbone"
         ) as mock_mera_simple:
             with mock.patch(
                 "piqture.gates.two_qubit_unitary.TwoQubitUnitary.simple_parameterization"
@@ -204,7 +204,7 @@ class TestMERA:
         # pylint: disable=line-too-long
         """Tests the mera_backbone method call via the mera_general function."""
         with mock.patch(
-            "piqture.tensor_network_circuits.mera.MERA.mera_backbone"
+            "piqture.tensor_networks.mera.MERA.mera_backbone"
         ) as mock_mera_general:
             with mock.patch(
                 "piqture.gates.two_qubit_unitary.TwoQubitUnitary.general_parameterization"
