+++ tests/tensor_networks/test_ttn.py
@@ -14,7 +14,7 @@ from __future__ import annotations
 from unittest import mock
 import pytest
 from qiskit.circuit import QuantumCircuit
-from piqture.tensor_network_circuits.ttn import TTN
+from piqture.tensor_networks.ttn import TTN
 from piqture.gates.two_qubit_unitary import TwoQubitUnitary
 
 
@@ -99,7 +99,7 @@ class TestTTN:
         # pylint: disable=line-too-long
         """Tests the ttn_backbone method call via the ttn_simple function."""
         with mock.patch(
-            "piqture.tensor_network_circuits.ttn.TTN.ttn_backbone"
+            "piqture.tensor_networks.ttn.TTN.ttn_backbone"
         ) as mock_ttn_simple:
             with mock.patch(
                 "piqture.gates.two_qubit_unitary.TwoQubitUnitary.simple_parameterization"
@@ -114,7 +114,7 @@ class TestTTN:
         # pylint: disable=line-too-long
         """Tests the ttn_backbone method call via the ttn_general function."""
         with mock.patch(
-            "piqture.tensor_network_circuits.ttn.TTN.ttn_backbone"
+            "piqture.tensor_networks.ttn.TTN.ttn_backbone"
         ) as mock_ttn_general:
             with mock.patch(
                 "piqture.gates.two_qubit_unitary.TwoQubitUnitary.general_parameterization"
